fix:修正 User 功能返回值為 Long。

+++ src/main/java/com/example/sunnymarketbackend/dao/UserDao.java
@@ -5,9 +5,9 @@ import com.example.sunnymarketbackend.entity.User;
 
 public interface UserDao {
 
-    User getUserById(Integer userId);
+    User getUserById(Long userId);
 
     User getUserByEmail(String email);
 
-    Integer createUser(UserRegisterRequest userRegisterRequest);
+    Long createUser(UserRegisterRequest userRegisterRequest);
 }
